Neovide

+++ os/ubuntu/setup.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
+
+absolute_here="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+here="$absolute_here"
+root="$here/../.."
 #
-echo "Symlinking ~/Application/nvim.appimage to /usr/bin/nvim . Will need sudo."
-sudo ln -s ~/Applications/nvim.appimage /usr/bin/nvim
+
+if [ -f "$HOME/Applications/nvim.appimage" ]; then
+  if [ ! -f "/usr/bin/nvim" ]; then
+    echo "Symlinking ~/Application/nvim.appimage to /usr/bin/nvim . Will need sudo."
+    sudo ln -s ~/Applications/nvim.appimage /usr/bin/nvim
+  fi
+fi
 
 ###############################################
 # Load Ubuntu/Gnome Configuration + KeyMappings
@@ -12,4 +21,9 @@ source "$PWD/os/ubuntu/configs.sh"
 # NOTE: to export dconf settings:
 # dconf dump / > ~/.config/dconf/user.d/everything.conf
 
+if [ -f "$HOME/Applications/neovide.AppImage" ]; then
+  cd "$HOME/Applications"
+  echo $PWD
+  appimage-to-desktop "neovide" "$HOME/Applications/neovide"
+fi
